==> project/0. Preparation.ipynb <==
<center>

# Preparation File

<center>

# Importations.
import zipfile
import gzip
import shutil
import os
import json
import pandas as pd
import ast

<div style="text-align: justify">

### 1. Unzipping the JSON files to do the project 

This code allowed me to copy and paste the zip file, unzip it, leaving the folder with the final files, and deleting all the zip files (the main and the other 3 zip files inside).

</div>

# Path to the zip file.
zip_file_path = 'PI MLOps - STEAM-20231219T030929Z-001.zip'

# Extraction directory.
destination_directory = ''

# Extract the .zip file.
with zipfile.ZipFile(zip_file_path, 'r') as zip_ref:
    zip_ref.extractall(destination_directory)

# Remove the zip file after extraction.
os.remove(zip_file_path)

# Folder containing the compressed .gz files.
compressed_folder = 'PI MLOps - STEAM'

# Extraction directory
destination_directory = 'PI MLOps - STEAM'

# Get the list of compressed .gz files in the folder.
compressed_files = [file for file in os.listdir(compressed_folder) if file.endswith('.gz')]

# Iterate over the list of files, decompress, and delete each one.
for compressed_file in compressed_files:
    # Full path of the compressed file
    compressed_file_path = os.path.join(compressed_folder, compressed_file)
    # Full path of the decompressed file
    decompressed_file_path = os.path.join(destination_directory, compressed_file.replace('.gz', ''))

    # Decompress the .gz file.
    with gzip.open(compressed_file_path, 'rb') as f_in:
        with open(decompressed_file_path, 'wb') as f_out:
            shutil.copyfileobj(f_in, f_out)

    print(f'File {compressed_file} decompressed to {decompressed_file_path}')

    # Remove the compressed file after decompression.
    os.remove(compressed_file_path)
    print(f'Compressed file {compressed_file} deleted.')

<div style="text-align: justify">

### 2. Converting Data: From JSON Format to CSV Format

These are the initial versions of the codes I used to attempt to convert the data and work with the files as data frames.

</div>

# For users_items.csv.
# File paths.
users_items = 'PI MLOps - STEAM/users_items.json'
users_items_csv = 'csv/users_items.csv'

# If the file does not exist, create the file.
if not os.path.exists(users_items_csv):
    # reading the Json data file.
    items = []
    with open(users_items, encoding='utf-8') as f:
        for line in f.readlines():
            items.append(ast.literal_eval(line))

    # transform the Json file into a dataframe.
    df_items = pd.DataFrame(items)

    # save the dataframe as a CSV file in the csv folder.
    df_items.to_csv(users_items_csv, index=False)
    print(f'The file {users_items_csv} was successfully created.')
else:
    print(f'The file {users_items_csv} already exists.')

The final version is inside the ETL_users_items.ipynb file.

# For user_reviews.csv.
# File paths.
user_reviews = 'PI MLOps - STEAM/user_reviews.json'
user_reviews_csv = 'csv/user_reviews.csv'

# If the file does not exist, create the file.
if not os.path.exists(user_reviews_csv):
    # reading the Json data file.
    reviews = []
    with open(user_reviews, encoding='utf-8') as f:
        for line in f.readlines():
            reviews.append(ast.literal_eval(line))

    # transform the Json file into a dataframe.
    #df_reviews = pd.DataFrame(reviews).
    df_reviews = pd.json_normalize(reviews, record_path=['reviews'], meta=['user_id', 'user_url'])

    # save the dataframe as a CSV file in the csv folder.
    df_reviews.to_csv(user_reviews_csv, index=False)
    print(f'The file {user_reviews_csv} was successfully created.')
else:
    print(f'The file {user_reviews_csv} already exists.')

The final version is inside the ETL_user_reviews.ipynb file.

# For steam_games.csv.
# File paths.
steam_games = 'PI MLOps - STEAM/steam_games.json'
steam_games_csv = 'csv/steam_games.csv'

# If the file does not exist, create the file.
if not os.path.exists(steam_games_csv):
    # reading the Json data file.
    games = []
    with open(steam_games, 'r') as f:
        # Iterates over the lines in the file and loads each line as a JSON object.
        for line in f:
            json_object = json.loads(line)
            games.append(json_object)

    # transform the Json file into a dataframe.
    df_games = pd.DataFrame(games)

    # save the dataframe as a CSV file in the csv folder.
    df_games.to_csv(steam_games_csv, index=False)
    print(f'The file {steam_games_csv} was successfully created.')
else:
    print(f'The file {steam_games_csv} already exists.')
