==> project/1. ETL_users_items.ipynb <==
<center>

# Users Items ETL

<center>

# Importations.
import os
import ast
import pandas as pd

<div style="text-align: justify">

### 1. From Json to csv 

I transformed the JSON file into a CSV file to read it as a data frame. After the first look, I checked that the column **items** was nested, so I had to unnest the column. I deleted the initial version, updated the code, and then created a new file to save space, as the file size was 560 MB, and the free limit for GitHub LFS is 1 GB.

</div>

# For users_items.csv.
# File paths.
users_items = 'PI MLOps - STEAM/users_items.json'
users_items_csv = 'csv/users_items.csv'

# If the file does not exist, create the file.
if not os.path.exists(users_items_csv):
    # reading the Json data file.
    items = []
    with open(users_items, encoding='utf-8') as f:
        for line in f.readlines():
            items.append(ast.literal_eval(line))
            
    # Transform the Json file into a DataFrame and normalize the 'items' column.
    df_items = pd.json_normalize(items, record_path=['items'], meta=['user_id', 'items_count', 'steam_id', 'user_url'])
            
    # Save the dataframe as a CSV file in the csv folder.
    df_items.to_csv(users_items_csv, index=False)
    print(f'The file {users_items_csv} was successfully created.')
else:
    print(f'The file {users_items_csv} already exists.')

# Reading the csv file.
df_items = pd.read_csv('csv/users_items.csv')

# First check to the file.
df_items.head(1)

The column **items** was nested; the composition of this column was a list that contained a dictionary.

# Reading the csv after the update.
df_items_unnested = pd.read_csv('csv/users_items.csv')

# Checking the columns and file content after the update.
df_items_unnested.head(3)

<div style="text-align: justify">

After normalization, the column **items** was unnested, and every game inside the **item_name** column was aligned with the correct user in the **user_id** column.

</div>

<div style="text-align: justify">

### 2. Checking NAN, None and Duplicate Items

I used 2 methods to find out if the file had NAN/None data.
but the results were negative, after applied the loc method I found that the file had 59104 duplicates.

</div>

# Procee to check if there is null data.
null_data = df_items_unnested.isnull().sum()
null_data

I used the isnull().sum() method, I could not find any null data.

# Second method to confirm if the file has any null data.
print(df_items_unnested.isna().any())

I tried a second variant to confirm the initial result.

# Finding duplicates.
duplicates = df_items_unnested.loc[df_items_unnested.duplicated()]
duplicates

After applied the **loc** method and the **duplicate** function, I found **59.104** duplicates.

# Total rows before deduplication.
total_rows_before = len(df_items_unnested)

# Remove duplicate rows.
df_items_unnested = df_items_unnested.drop_duplicates(keep='first')

# Total rows after deduplication.
total_rows_after = len(df_items_unnested)

# Total rows removed.
rows_removed = total_rows_before - total_rows_after

# Print the information.
print(f'Total rows before: {total_rows_before}')
print(f'Total rows after: {total_rows_after}')
print(f'Rows removed: {rows_removed}')

I wanted to verify if the **drop_duplicates** function was going to delete the total amount of duplicates.

<div style="text-align: justify">

### 3. Deleting Columns

I deleted some columns that I considered unnecessary for the project. This helped me reduce the size of the file. The columns I removed were **playtime_2weeks** (which contained the same information as playtime forever), **steam_id** (which duplicated the information found in user_id), **items_count** (representing the number of items/games per user), and **user_url** (which contained the link to the user's profile).

</div>

# Deleting the columns.
df_items_unnested = df_items_unnested.drop(['playtime_2weeks', 'items_count', 'steam_id', 'user_url'], axis=1)

<div style="text-align: justify">

### 4. Checking the Playtime Forever Column

For the project, it was not necessary to retain information about items/games with 0 minutes of playtime. Steam does not allow you to write a review unless you have started/played the game.

</div>

# Filtrar las filas donde playtime_forever es menor de 30 minutos
df_items_unnested = df_items_unnested[df_items_unnested['playtime_forever'] >= 300]

# checking the result.
df_sorted = df_items_unnested.sort_values(by='playtime_forever')
df_sorted

After deleting the duplicates, the file initially had 5,094,105 rows, and now it only has 3,246,375. This means that there were 1,847,730 rows with 0 minutes.
Update: I had to further reduce the size of this file because the processing time for graphics was too high.

df_items_unnested.info()

# resetting the index of the file.
df_items_unnested.reset_index(drop=True, inplace=True)
df_items_unnested

I wanted to reset the index to have a better order and future reference if I need it.

<div style="text-align: justify">

### 5. It is time to overwrite it

The file size was reduced to just 137 MB.

</div>

# Overwrite the original CSV file.
df_items_unnested.to_csv('csv/users_items.csv', index=False)
