==> project/5. Sentiment_analysis.ipynb <==
<center>

# Sentiment Analysis

<center>

<div style="text-align: justify">

I used TextBlob because it can assess the polarity (positive, negative, or neutral) and subjectivity (objective or subjective) of a text. I explored different thresholds and decided to use 0.3 because, in my opinion, it showed the most balanced results.

</div>

# Importantions.
import pandas as pd
from textblob import TextBlob

df_sentiment = pd.read_csv('csv/user_reviews.csv')

# Sentiment Analysis Function.
def sentiment_analysis(text):
    if pd.notna(text):
        blob = TextBlob(text)
        sentiment = blob.sentiment.polarity
        if sentiment < -0.3:
            return 0  # negative
        elif sentiment > 0.3:
            return 2  # positive
        else:
            return 1  # neutral
    else:
        return 1  # missing

# Apply and create the sentiment_analysis column.
df_sentiment['sentiment_analysis'] = df_sentiment['review'].apply(sentiment_analysis)

# Calculate the total count of positive, negative, and neutral responses
positive_count = df_sentiment[df_sentiment['sentiment_analysis'] == 2].shape[0]
negative_count = df_sentiment[df_sentiment['sentiment_analysis'] == 0].shape[0]
neutral_count = df_sentiment[df_sentiment['sentiment_analysis'] == 1].shape[0]

# Print the results
print(f"Positive Responses: {positive_count}")
print(f"Negative Responses: {negative_count}")
print(f"Neutral Responses: {neutral_count}")

# Drop the 'review' column.
df_sentiment = df_sentiment.drop('review', axis=1)

df_sentiment.head(1)

# Saving a new file.
df_sentiment.to_csv('datasets/user_reviews_SA.csv', index=False)
