==> project/6. API_queries.ipynb <==
<center>

# API Queries

<center>

In this file, I prepared all the datasets required for the functions in main.py.

# Importations.
import pandas as pd

# Reading CSV files.
df_items = pd.read_csv('csv/users_items.csv')  
df_reviews = pd.read_csv('csv/user_reviews.csv')
df_games = pd.read_csv('csv/steam_games.csv')    

# Reading Datasets files.
df_function_1_2 =  pd.read_csv('datasets/function_1_2.csv')
df_reviews_SA =  pd.read_csv('datasets/user_reviews_SA.csv')
df_function_3_4_5 = pd.read_csv('datasets/function_3_4_5.csv')
df_function_item_item = pd.read_csv('datasets/function_item_item.csv')

df_games.head(1)

df_reviews_SA.head(1)

<div style="text-align: justify">

### 1. Dataset for the PlayTimeGenre and UserForGenre functions

For the first and second functions, I required:

item_id, playtime_forever, and user_id from the file csv/users_items.csv.

item_id, genres, and year from the file csv/steam_games.csv.

</div>

# Select the columns you want to combine
columns_items = df_items[['item_id', 'playtime_forever', 'user_id']]
columns_games = df_games[['item_id', 'genres', 'year']]

# Combine the columns into a new DataFrame
df_function_1_2 = pd.merge(columns_items, columns_games, on='item_id', how='inner')

# Convert 'genres' column to lowercase
df_function_1_2['genres'] = df_function_1_2['genres'].str.lower()

# Save the new DataFrame to a CSV file
df_function_1_2.to_csv('datasets/function_1_2.csv', index=False)

df_function_1_2.head(1)

<div style="text-align: justify">

### 2. Dataset for the UsersRecommend, UsersWorstDeveloper and sentiment_analysis functions

For the third, fourth, and fifth functions. For this file, I needed:

item_id, sentiment_analysis, and recommend from the file datasets/user_reviews_SA.csv.

item_id, title, year, and Developer from the file csv/steam_games.csv.

</div>

# Select the necessary columns for each DataFrame
columns_user_reviews = df_reviews_SA[['item_id', 'sentiment_analysis', 'recommend']]
columns_steam_games = df_games[['item_id', 'title', 'year', 'developer']]

# Perform Inner Join
df_function_3_4_5 = pd.merge(columns_user_reviews, columns_steam_games, on='item_id', how='inner')

# Set all values to 1, representing negative sentiment.
df_function_3_4_5['reviews_recommend'] = 1

# When the condition is true (1 or 2), set the value to 2, representing positive sentiment.
condition = (df_function_3_4_5['recommend'] == True) & (df_function_3_4_5['sentiment_analysis'].isin([1, 2]))
df_function_3_4_5.loc[condition, 'reviews_recommend'] = 2

# Convert 'developer' column to lowercase
df_function_3_4_5['developer'] = df_function_3_4_5['developer'].str.lower()

# Reorganize the columns.
column_order = ['item_id', 'title', 'recommend', 'developer', 'year', 'sentiment_analysis', 'reviews_recommend']
df_function_3_4_5 = df_function_3_4_5[column_order]

# Save the result to a new CSV file
df_function_3_4_5.to_csv('datasets/function_3_4_5.csv', index=False)

df_function_3_4_5.head(1)

<div style="text-align: justify">

### 3. Dataset for the recommendation_game(product_id).

For the first recommendation model, I needed:

item_id, sentiment_analysis and recommend from the file datasets/user_reviews_SA.

item_id, title, genres from the file csv/steam_games.csv.

</div>

# Select the necessary columns for each DataFrame.
columns_user_reviews = df_reviews_SA[['item_id', 'sentiment_analysis', 'recommend']]
columns_steam_games = df_games[['item_id', 'title', 'genres']]

# Suponiendo que df_function_item_item es tu DataFrame
df_function_item_item = pd.merge(columns_user_reviews, columns_steam_games, on='item_id', how='inner')

# Set all values to 1, representing negative sentiment.
df_function_item_item['reviews_rating'] = 0

# When the condition is true (1 or 2), set the value to 2, representing positive sentiment.
condition = (df_function_item_item['recommend'] == True) & (df_function_item_item['sentiment_analysis'].isin([1, 2]))
df_function_item_item.loc[condition, 'reviews_rating'] = 1

# Drop duplicates based on all columns
# df_function_item_item = df_function_item_item.drop_duplicates()

# Drop columns.
df_function_item_item = df_function_item_item.drop(['sentiment_analysis', 'recommend'], axis=1)

# Save the result to a new CSV file.
df_function_item_item.to_csv('datasets/function_item_item.csv', index=False)

df_function_item_item.head(1)
